// web/src/core/types.ts
export enum TokenType {
    // Keywords
    HI_KAKA = "HI_KAKA",
    BYE_KAKA = "BYE_KAKA",
    VAR_KAKA = "VAR_KAKA",
    PRINT_KAKA = "PRINT_KAKA",
    IF_KAKA = "IF_KAKA",
    ELSE_KAKA = "ELSE_KAKA",
    WHILE_KAKA = "WHILE_KAKA",
    BREAK_KAKA = "BREAK_KAKA",
    CONTINUE_KAKA = "CONTINUE_KAKA",
    TRUE_KAKA = "TRUE_KAKA",
    FALSE_KAKA = "FALSE_KAKA",
    NULL_KAKA = "NULL_KAKA",
    INPUT_KAKA = "INPUT_KAKA",
    AND_KAKA = "AND_KAKA",
    OR_KAKA = "OR_KAKA",
    FUNCTION_KAKA = "FUNCTION_KAKA",
    RETURN_KAKA = "RETURN_KAKA",
    OPEN_BRACKET = "OPEN_BRACKET",
    CLOSE_BRACKET = "CLOSE_BRACKET",

    // Data Types
    NUMBER = "NUMBER",
    STRING = "STRING",
    IDENTIFIER = "IDENTIFIER",

    // Operators
    EQUALS = "EQUALS",
    PLUS = "PLUS",
    MINUS = "MINUS",
    MULTIPLY = "MULTIPLY",
    DIVIDE = "DIVIDE",
    MODULO = "MODULO",

    // Logical/Comparison
    EE = "EE", // ==
    NE = "NE", // !=
    GT = "GT", // >
    LT = "LT", // <
    GTE = "GTE", // >=
    LTE = "LTE", // <=

    // Punctuation
    SEMICOLON = "SEMICOLON",
    OPEN_BRACE = "OPEN_BRACE",
    CLOSE_BRACE = "CLOSE_BRACE",
    OPEN_PAREN = "OPEN_PAREN",
    CLOSE_PAREN = "CLOSE_PAREN",
    COMMA = "COMMA",
    EOF = "EOF"
}

export interface Token {
    type: TokenType;
    value: string;
    line: number;
}